Remove unused files and refactor
