+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Crypto Stat",
+  "name": "Crypto Stat",
   "start_url": ".",
   "display": "standalone",
   "theme_color": "#000000",
