--- src/actions/displayActions.js
@@ -1,9 +1,0 @@
-const darkMode = () => ({
-  type: 'DARK_MODE_ON',
-});
-
-const lightMode = () => ({
-  type: 'DARK_MODE_OFF',
-});
-
-export { darkMode, lightMode };
