--- src/actions/index.js
@@ -1,6 +1,0 @@
-import { lightMode, darkMode } from './displayActions';
-import { updateCryptos, filterCryptos, toggleAscSorting } from './cryptosActions';
-
-export {
-  lightMode, darkMode, updateCryptos, filterCryptos, toggleAscSorting,
-};
