+++ src/app/component/_tests_/CryptoPage.test.js
