+++ src/app/component/_tests_/CryptoTableData.test.js
