+++ src/app/component/helpers/MoneyFormat.js
