+++ src/app/component/helpers/NumberFilters.js
