--- src/css/chart.css
@@ -1,32 +1,0 @@
-.chart {
-  width: 100%;
-  font-size: 0.9rem;
-}
-
-.chart-header {
-  border-bottom: 1px solid grey;
-  border-top: 1px solid grey;
-  font-weight: bold;
-}
-
-.chart button {
-  border: 0;
-  background-color: inherit;
-  color: inherit;
-}
-
-.chart-header td {
-  padding: 15px 0;
-}
-
-.chart-header td button {
-  padding: 0;
-}
-
-.chart-crypto-row {
-  border-bottom: 1px solid grey;
-}
-
-.chart-crypto-row > * {
-  padding: 7px 0;
-}
