--- src/css/hotToday.css
@@ -1,7 +1,0 @@
-.today-view {
-  box-shadow: rgba(100, 100, 111, 0.2) 0 7px 29px 0;
-}
-
-.today-view span {
-  font-weight: bold;
-}
