+++ src/pages/Cryptocurrencies/Component/DetailsDrawer.js
