+++ src/pages/_tests_/CryptoTable.test.js
