--- src/reducers/darkmode.js
@@ -1,20 +1,0 @@
-const darkModeReducer = (state = [], action) => {
-  switch (action.type) {
-    case 'DARK_MODE_ON':
-      return {
-        ...state,
-        darkmode: true,
-      };
-
-    case 'DARK_MODE_OFF':
-      return {
-        ...state,
-        darkmode: false,
-      };
-
-    default:
-      return state;
-  }
-};
-
-export default darkModeReducer;
